Guard against over-reading and strange offsets
With strangely formatted text entries, we can sometimes overread what we are supposed to, this commit fixes that by verifying we didn't overread and, in the case we do, redo the processing by reading byte by byte and sending combined unicode chars after every 2 bytes read.
The second part of this commit guards against strange offsets, usually pointing to data inside the header. We now only create blank entries for them, and setup new JSON values to notify the repacker when an entry should be ignored and when it has its own defined offset.