Script Handling Improvements
Do not read into the string offsets section when checking if we are done with the script attributes.
Do not parse "{" and "}" as ASCII tags if we are in the middle of reading an id tag
HACK: If the text content buffer length mod 0x10 = 0, do not pad to 0x20.