Complete Refactor + Version 2 Script Extraction
A new script version was discovered, it is more similar to the old version but has a more struct like format in the string offset section. This required a complete refactor of the script extraction process. There is now a single function that can handle multiple codepaths depending on an argument given to it at runtime. Version 2 scripts are only capable of extracting at this time.