Remove old code, V2 Script Repacking
There is a hack with v2 script repacking where the data offset for the string offset offsets section is assumed to be 60 bytes. This might have to be changed in the future. The tdmfscript.d file should also be easier to navigate now that its just the main two functions in there.